requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
black==21.12b0
cffi==1.15.0
channels==3.0.4
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1       
daphne==3.0.2
Django==3.2.5
django-cors-headers==3.7.0 
djangorestframework==3.12.4
geographiclib==1.52        
geopy==2.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.1
platformdirs==2.4.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pytz==2021.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
tomli==1.2.3
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==4.0.1
zope.interface==5.4.0